+++ braindecode/models/__init__.py
@@ -2,6 +2,7 @@
 Some predefined network architectures for EEG decoding.
 """
 from .base import EEGModuleMixin
+from .biot import BIOT
 from .eegconformer import EEGConformer
 from .eegitnet import EEGITNet
 from .deep4 import Deep4Net
@@ -22,8 +23,8 @@ from .tidnet import TIDNet
 from .usleep import USleep
 from .util import get_output_shape, to_dense_prediction_model
 from .modules import TimeDistributed
-
-from .util import _init_models_dict
+from .util import _init_models_dict, models_mandatory_parameters
+from .labram import Labram
 
 # Call this last in order to make sure the dataset list is populated with
 # the models imported in this file.